Adding more comments and changing intake ports
Added some comments to the various controller buttons for ease of use.
Also set right intake to port 2 as that is what is wired to now. Left
intake is not yet wired but when it is it will be port three. The only
reason intake was not working was due to the motors not being wired yet.