Turned off (disabled) Gyro for driving practices
This is a quick shortcut of Gyro to get the practice running. Not the
Gyro itself was causing problems but more likely the PID controller. I'm
committing this change just as a checkpoint at which the bot was running
at the practice field. Should be re-enabled for further Gyro research.