Changed MooseLifter to new Design specs
We are no longer using motors and limit switches for the moose lifter.
We now have 2 solenoids. One for lifting and lowering, and one to
prevent the other solenoid from going all the way. MooseLifter will only
ever need to go all the way up and all the way down, so floats are
changed to booleans.