Added 2nd motor to Lifter
Lifter design change: Now has 2 motors. One for lifting and one for
pushing stacks out. There will be two limit switches preventing the new
motor from going all the way out and in. Added ability to push in
ControlLifter so we can move up and down while moving in and out if we
need to. Also changed ports of LIFTER and PUSHER in RobotMap because
Drive Team wants the lifter to be controlled by the weapons officer, so
I moved them to the axes. Pusher will not have encoder.