Removed last year's nonrelevant code
Delete rectangle count and sorting since the Roborealm returns only
variable information about the Center of Gravity. Center of Gravity
returns a square box on roborealm, so there is no need for finding the
rectangle height and width, since it is a guaranteed square. When
building, lots of warnings about unused variables, but no errors I
think.