Cleaned code and added smartdashboard access
Added smartdashboard access to lifter and pusher to figure out what's
wrong. Pusher now works. It was set to the wrong axis, 4 instead of 3,
since the drawing on the driver station is actually wrong, and I changed
it so that it sets the control mode every time pushLifter is called, not
just in initialization.