Pusher LED fixes
Switched orientation of Pusher CANTalon limit switches. Before, the LED
would only light up if the joystick was holding against the limit
switch, instead of just lighting up when the limit switch was pressed.
Function is a bit weird since it outputs an integer instead of a
boolean, but it should work. If the orientation is reversed I'll just
negate it and add 1.