Changes from field 3/24/2015
Reverted previous Pusher change due to difficulties, fix likely found
(view line notes on previous pusher change). Tuned 3Tote auton values to
work better with the new start angle of the robot, which is more easily
reproducible on the field. Commented out lines are commented out for a
reason, they may be needed but are not with the current configuration.