Two parallel lifter tasks and parallel concept
AutonomousGroupBincher is a task that can run parallel.
AutonomousGroupReload is a task that drops the stack on a tote and grabs
that tote instead and can run parallel.
AutonWaitForFinished is a tiny command that waits for a certain command
to finish (IsFinished is a protected method, so IsRunning is used
instead).
Another massive cleanup and renaming in Group8.